refactor xml normalization parser
